# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST para el sistema de monitoreo de instancias Proxmox El Centinela.\n**Flujo de autenticación**: Login → `/auth/login` → obtener `jwtTemporal` → `/auth/2fa/verify` → obtener `accessToken` → usar en el botón **Authorize** de esta UI.",
        "title": "El Centinela API",
        "contact": {
            "name": "Soporte El Centinela"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/instances": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lee en vivo el inventario de Proxmox VE (VMs y contenedores). Un ADMIN recibe el cluster completo; un OPERATOR recibe únicamente las instancias que tiene asignadas.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Instancias Proxmox"
                ],
                "summary": "Listar instancias",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ports.InstanciaListadaDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "INTERNAL_ERROR — error al consultar los permisos del usuario",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "PROXMOX_UNAVAILABLE",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "504": {
                        "description": "PROXMOX_UNAVAILABLE — Proxmox no respondió a tiempo",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/instances/{vmid}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Devuelve el estado actual (running/stopped, tipo, nodo) de una VM o contenedor leído en vivo desde Proxmox VE.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Instancias Proxmox"
                ],
                "summary": "Obtener instancia",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "VMID de la instancia",
                        "name": "vmid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ports.InstanciaProxmoxDTO"
                        }
                    },
                    "400": {
                        "description": "INVALID_VMID",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "INSTANCE_ACCESS_DENIED — el OPERATOR no tiene este vmid asignado",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "INSTANCE_NOT_FOUND",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "PROXMOX_UNAVAILABLE",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "504": {
                        "description": "PROXMOX_UNAVAILABLE — Proxmox no respondió a tiempo",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/instances/{vmid}/start": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Dispara el arranque de una VM o contenedor en Proxmox VE. La operación es asíncrona: devuelve el UPID de la tarea que Proxmox crea para seguir su progreso.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Instancias Proxmox"
                ],
                "summary": "Iniciar instancia",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "VMID de la instancia",
                        "name": "vmid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "upid de la tarea en Proxmox y tareaId (llega en el evento TASK_FINISHED al terminar)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "INVALID_VMID",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "INSTANCE_ACCESS_DENIED — el OPERATOR no tiene este vmid asignado",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "INSTANCE_NOT_FOUND",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "PROXMOX_UNAVAILABLE",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "504": {
                        "description": "PROXMOX_UNAVAILABLE — Proxmox no respondió a tiempo",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/instances/{vmid}/stop": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Dispara el apagado forzado de una VM o contenedor en Proxmox VE. La operación es asíncrona: devuelve el UPID de la tarea que Proxmox crea para seguir su progreso.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Instancias Proxmox"
                ],
                "summary": "Detener instancia",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "VMID de la instancia",
                        "name": "vmid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "upid de la tarea en Proxmox y tareaId (llega en el evento TASK_FINISHED al terminar)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "INVALID_VMID",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "INSTANCE_ACCESS_DENIED — el OPERATOR no tiene este vmid asignado; INSTANCE_PROTECTED — la instancia es infraestructura de El Centinela",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "INSTANCE_NOT_FOUND",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "PROXMOX_UNAVAILABLE",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "504": {
                        "description": "PROXMOX_UNAVAILABLE — Proxmox no respondió a tiempo",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "http.ErrorResponse": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "string",
                    "example": "INVALID_REQUEST"
                },
                "message": {
                    "type": "string",
                    "example": "Mensaje descriptivo del error."
                }
            }
        },
...
        "ports.InstanciaListadaDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "node": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "description": "\"vm\" | \"lxc\"",
                    "type": "string"
                }
            }
        },
        "ports.InstanciaProxmoxDTO": {
            "type": "object",
            "properties": {
                "estado": {
                    "description": "\"running\" | \"stopped\" | ...",
                    "type": "string"
                },
                "nodo": {
                    "type": "string"
                },
                "nombre": {
                    "type": "string"
                },
                "tipo": {
                    "description": "\"qemu\" | \"lxc\"",
                    "type": "string"
                },
                "vmid": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Token de acceso definitivo. Formato: **Bearer \u0026lt;accessToken\u0026gt;**. Obtené el token completando el flujo: `POST /auth/login` → `POST /auth/2fa/verify`.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BearerPreAuth": {
            "description": "JWT temporal pre-2FA. Formato: **Bearer \u0026lt;jwtTemporal\u0026gt;**. Obtené el token de `POST /auth/login`.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}